added login logic
