+++ app.js
@@ -2,8 +2,8 @@ require("dotenv").config();
 const express = require("express");
 const logger = require("morgan");
 const cors = require("cors");
-const contactsRouter = require("./routes/api/contacts");
-const userRouter = require('./routes/api/users');
+const contactsRouter = require("./routes/api/contactsRouter");
+const userRouter = require('./routes/api/usersRouter');
 const app = express();
 // const formatsLogger = app.get("env") === "development" ? "dev" : "short";
 const formatsLogger = process.env.NODE_ENV === "development" ? "dev" : "short";
@@ -13,7 +13,7 @@ app.use(cors());
 app.use(express.json());
 
 app.use("/api/contacts", contactsRouter);
-app.use("/auth/users", userRouter);
+app.use("/api/users", userRouter);
 
 app.use((req, res) => {
   res.status(404).json({ message: "Not found" });
