+++ utils/httpError/httpError.js
@@ -6,4 +6,14 @@ class httpError extends Error {
     }
 }
 
+// const errorHandler = (err, req, res, next) => {
+//     if (err instanceof httpError) {
+//         return res.status(err.status).json({message: err.message})
+//     }
+//     res.status(500).json({ message: err.message });
+// }
+
 module.exports = httpError;
+
+// module.exports = { httpError, errorHandler };
+
